app/pipeline/caption_generator.py
from transformers import BlipProcessor, BlipForConditionalGeneration
from PIL import Image

processor = BlipProcessor.from_pretrained("Salesforce/blip-image-captioning-base")
model = BlipForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-base")

def generate_captions(frames):
    results = []

    for f in frames:
        img = Image.open(f).convert('RGB')
        inputs = processor(img, return_tensors="pt")

        out = model.generate(**inputs)
        caption = processor.decode(out[0], skip_special_tokens=True)

        results.append(caption)

    return results