requirements.txt
torch
transformers
opencv-python
ffmpeg-python
openai-whisper
numpy
pillow
tqdm
requests
python-dotenv